+++ .storybook/config.js
@@ -2,6 +2,8 @@ import React from 'react';
 import { configure, addDecorator, addParameters } from '@storybook/react';
 import { setDefaults } from 'react-storybook-addon-props-combinations'
 import { withRootAttribute } from 'storybook-addon-root-attribute';
+import { Provider } from 'react-redux';
+import { combineReducers, createStore } from 'redux';
 import { Setup } from '../src';
 
 setDefaults({
@@ -16,6 +18,27 @@ function loadStories() {
   });
 }
 
+const initialState = {
+  theme: 'LIGHT',
+};
+
+function userReducer(state = initialState, { payload, type }) {
+  switch (type) {
+    default:
+      return state;
+  }
+}
+
+const store = createStore(combineReducers({
+  user: userReducer,
+}));
+
 addDecorator(withRootAttribute)
-addDecorator(storyFn => <Setup>{storyFn()}</Setup>);
+addDecorator(storyFn => {
+  return (
+    <Provider store={store}>
+      <Setup>{storyFn()}</Setup>
+    </Provider>
+  );
+});
 configure(loadStories, module);